third_party/WebKit/LayoutTests/editing/spelling/markers-expected.txt

Tests spelling markers for misspellings.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.markerCountForNode(element.firstChild, next.marker) became 1
PASS range.toString() is next.issue
PASS internals.markerCountForNode(element.firstChild, next.marker) became 1
PASS range.toString() is next.issue
PASS successfullyParsed is true

TEST COMPLETE
zz.
orange,zz,apple.
